move functions requiring the `distro` package to a separate file to prevent ImportError on run
